% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vDE.R
\name{vDE}
\alias{vDE}
\title{Differential gene expression analysis for all identity classes
Find differentially expressed genes in features for each clusters in a dataset}
\usage{
vDE(
  object = object,
  features.by = features.by,
  min.cells = 20,
  clusters = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{features.by}{A vector of features which is used to divid cells into two groups}

\item{min.cells}{Minimum number of cells to do DGE}

\item{clusters}{Clusters selected to do DGE}
}
\description{
Differential gene expression analysis for all identity classes
Find differentially expressed genes in features for each clusters in a dataset
}
\author{
rstatistics
}
