% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGO.R
\name{vGO}
\alias{vGO}
\title{vGO annotation}
\usage{
vGO(object = object, key = "vDE", cluster = cluster, feature.by = feature.by)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The DEGs are stored in object@misc$vDE$key}

\item{cluster}{Cluster specified by user}

\item{feature.by}{Feature which is used to divid cells into two groups}
}
\value{
a list object that stores vGO results
}
\description{
vGO
vGO annotation for specific cluster
}
\author{
rstatistics
}
