% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vGODotPlot.R
\name{vGODotPlot}
\alias{vGODotPlot}
\title{GODotPlot visualization}
\usage{
vGODotPlot(
  object = object,
  key = "vGO",
  cluster = cluster,
  feature.by = feature.by,
  ont = "BP",
  top_n = 20
)
}
\arguments{
\item{object}{A seurat object}

\item{key}{Key name. The vGO results are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{feature.by}{Feature which is used to divid cells into two groups}

\item{ont}{Ontology to use, only "BP", "CC", or "MF".}

\item{top_n}{Only top_n entries to plot}
}
\value{

}
\description{
vGODotPlot
vGO DotPlot visualization for specific cluster
}
\author{
rstatistics
}
