% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vHeatMap.R
\name{vHeatMap}
\alias{vHeatMap}
\title{vHeatMap visualization}
\usage{
vHeatMap(
  object = object,
  key = "vDE",
  cluster = cluster,
  feature.by = feature.by,
  markers = NULL,
  colors = NULL,
  top_n = 5,
  font.size = 8,
  assay = NULL,
  slot = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The DE matrix are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{feature.by}{Feature which is used to divid cells into two groups}

\item{markers}{A vector of features to plot, default is top_n features}

\item{colors}{User defined colours}

\item{top_n}{Top n markers to plot}

\item{font.size}{Font size}

\item{assay}{Assay to pull from}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}
}
\value{

}
\description{
Average HeatMap for specific feature
Average HeatMap for specific feature in all clusters
}
\author{
rstatistics
}
