% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vReadTenX.R
\name{vReadTenX}
\alias{vReadTenX}
\title{xReadTenX}
\usage{
vReadTenX(
  sample = sample,
  path = path,
  column = column,
  project = project,
  group = group,
  min.nFeature = min.nFeature,
  max.nFeature = max.nFeature,
  percent.mito = percent.mito,
  min.nCount_RNA = min.nCount_RNA,
  max.nCount_RNA = max.nCount_RNA
)
}
\arguments{
\item{sample}{Tab-delimited sample information file, such as sample, path}

\item{path}{Path that holds scRNA matrix}

\item{column}{Column of genes.tsv or features.tsv to use for gene names, default column=2}

\item{project}{Project name for the Seurat object}

\item{group}{Group name for the Seurat object}

\item{min.nFeature}{Minimum number of features in a cell}

\item{max.nFeature}{Maximum number of features in a cell}

\item{percent.mito}{Maximum percent of mito in a cell}

\item{min.nCount_RNA}{Minimum number of UMIs in a cell}

\item{max.nCount_RNA}{Maximum number of UMIs in a cell}
}
\value{
Seurat object.
}
\description{
vReadTenX
vReadTenX reads host and pathogen matrix simultaneously
}
\author{
rstatistics
}
