% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vVolcanoPlot.R
\name{vVolcanoPlot}
\alias{vVolcanoPlot}
\title{VolcanoPlot visualization}
\usage{
vVolcanoPlot(
  object = object,
  key = "vDE",
  cluster = cluster,
  feature.by = feature.by,
  top_n = NULL
)
}
\arguments{
\item{object}{A seurat object.}

\item{key}{The DE matrix are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{feature.by}{Feature which is used to divid cells into two groups}

\item{top_n}{Number of top up/down genes to show.}
}
\description{
VolcanoPlot for specific feature
}
\author{
rstatistics
}
