% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xDE.R
\name{xDE}
\alias{xDE}
\title{Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters in a dataset}
\usage{
xDE(
  object = object,
  comparison = c("condA", "condB"),
  group.by = NULL,
  min.cells = 20,
  clusters = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{comparison}{Vector of nominal variables from group.by. Eg., comparison=c("Normal", "Tumor")}

\item{group.by}{Regroup cells before performing DE}

\item{min.cells}{Minimum number of cells to perform DE}

\item{clusters}{Vector of Clusters to perform DE}

\item{return}{A list.}
}
\description{
Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters in a dataset
}
\author{
rstatistics
}
