% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xFeaturePlot.R
\name{xFeaturePlot}
\alias{xFeaturePlot}
\title{xFeaturePlot}
\usage{
xFeaturePlot(
  object = object,
  reduction = NULL,
  features = features,
  pt.size = NULL,
  font.size = NULL,
  cols = NULL,
  ncol = NULL,
  combine = FALSE,
  order = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction to use}

\item{features}{Vector of features to plot}

\item{pt.size}{Point size to plot}

\item{font.size}{Font size}

\item{cols}{Vector of two colors to form the gradient over}

\item{combine}{Combine plots into a single patchworked ggplot object}

\item{order}{Determine whether to plot cells in order of expression}

\item{ncols}{Number of columns to combine features}
}
\value{

}
\description{
xFeaturePlot
wrapper for Seurat::FeaturePlot
}
\author{
rstatistics
}
