% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGO.R
\name{xGO}
\alias{xGO}
\title{GO annotation}
\usage{
xGO(object = object, key = "DE", clusters = NULL, only.pos = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The DEGs are stored in object@misc$key}

\item{clusters}{A vector of clusters, default clusters=NULL}

\item{only.pos}{Only use Upregulated genes. Default only.pos=TRUE}
}
\value{
a list object that stored GO results.
}
\description{
Cluster GO annotation
GO annotation for each cluster
}
\author{
rstatistics
}
