% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGOBarPlot.R
\name{xGOBarPlot}
\alias{xGOBarPlot}
\title{GOBarPlot visualization}
\usage{
xGOBarPlot(
  object = object,
  key = "GO",
  cluster = cluster,
  ont = "BP",
  top_n = 20
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The vGO results are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{ont}{Ontology to use, only "BP", "CC", or "MF".}

\item{top_n}{Only top_n entries to plot}
}
\value{

}
\description{
GO BarPlot visualization
GO BarPlot visualization for each cluster
}
\author{
rstatistics
}
