% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGOBoxPlot.R
\name{xGOBoxPlot}
\alias{xGOBoxPlot}
\title{GOBoxPlot visualization}
\usage{
xGOBoxPlot(object = object, key = "GO", ont = "BP", top_n = 5)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The DEGs are stored in object@misc$key}

\item{ont}{Ontology to use, only "BP", "CC", and "MF".}

\item{top_n}{Only top_n entries to plot.}
}
\value{

}
\description{
GO BoxPlot visualization
GO BoxPlot visualization for all clusters
}
\author{
rstatistics
}
