% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIntegrate.R
\name{xIntegrate}
\alias{xIntegrate}
\title{xIntegrate}
\usage{
xIntegrate(
  x = "/home/scrna/Lung/Pathogen-Track/scNSCLC/SampleInfo.txt",
  min.nFeature = 500,
  max.nFeature = 4000,
  percent.mito = 10,
  min.nCount_RNA = 1000,
  max.nCount_RNA = Inf,
  nVariable = 2000,
  nPC = 30,
  res = 0.5
)
}
\arguments{
\item{x}{Tab-delimited sample information file, such as sample, path}

\item{min.nFeature}{Minimum number of features in a cell}

\item{max.nFeature}{Maximum number of features in a cell}

\item{percent.mito}{Maximum percent of mito in a cell}

\item{min.nCount_RNA}{Minimum number of UMIs in a cell}

\item{max.nCount_RNA}{Maximum number of UMIs in a cell}

\item{nVariable}{Number of features to select as top variable features}

\item{nPC}{Total Number of PCs to compute}

\item{res}{Resolution parameter to set. Default res=0.5}

\item{path}{Path that holds scRNA matrix}

\item{column}{Column of genes.tsv or features.tsv to use for gene names, default column=2}

\item{project}{Project name for the Seurat object}

\item{group}{Group name for the Seurat object}
}
\value{
Seurat object.
}
\description{
xIntegrate Matrix
xIntegrate scRNA-seq
}
\author{
rstatistics
}
