% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPopulationPlot.R
\name{xPopulationPlot}
\alias{xPopulationPlot}
\title{Population Plot
A stacked bar plot to show the fraction composition of each sample or cluster}
\usage{
xPopulationPlot(object, by, cols)
}
\arguments{
\item{object}{Seurat object}

\item{by}{String used to separate cluster/sample, only "cluster" or "sample" is accepted}

\item{cols}{Vector of colors, each color corresponds to an identity class}
}
\value{

}
\description{
Population Plot
A stacked bar plot to show the fraction composition of each sample or cluster
}
