% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xReadTenX.R
\name{xReadTenX}
\alias{xReadTenX}
\title{xReadTenX}
\usage{
xReadTenX(
  sample = sample,
  path = path,
  column = column,
  project = project,
  group = group,
  min.nFeature = min.nFeature,
  max.nFeature = max.nFeature,
  percent.mito = percent.mito,
  nCount = nCount
)
}
\arguments{
\item{sample}{Tab-delimited sample information file, such as sample, path}

\item{path}{Minimum number of features in a cell.}

\item{column}{Maximum number of features in a cell.}

\item{project}{Maximum percent of mito in a cell.}

\item{group}{Number of variable features.}

\item{min.nFeature}{Minimum number of features in a cell.}

\item{max.nFeature}{Maximum number of features in a cell.}

\item{percent.mito}{Maximum percent of mito in a cell.}

\item{nCount_RNA}{Minimum number of UMIs in a cell.}
}
\value{
a seurat object.
}
\description{
xReadTenX
xReadTenX reads host and pathogen matrix simultaneously
}
\author{
rstatistics
}
