% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVolcanoPlot.R
\name{xVolcanoPlot}
\alias{xVolcanoPlot}
\title{VolcanoPlot visualization}
\usage{
xVolcanoPlot(object = object, key = key, cluster = cluster, top_n = NULL)
}
\arguments{
\item{object}{A seurat object.}

\item{key}{Key name. The DE matrix are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{top_n}{Number of top up/down genes to show.}
}
\description{
VolcanoPlot for specific group
}
\author{
rstatistics
}
